changed: merged cleaup branch (mainly format/signed-unsigned)
git-svn-id: https://opensg.vrsource.org/svn/trunk@2086 4683daeb-ad0f-0410-a623-93161e962ae5

+++ Source/Contrib/ComplexSceneManager/Helper/OSGCSMKeySensorHelper.cpp
@@ -83,7 +83,7 @@ void CSMKeySensorHelper::updateSensors(KeySensor *pSensor)
 
         if(kIt != _oKeyMap[i].end())
         {
-            if(pSensor->getKey() != i)
+            if(pSensor->getKey() != Int32(i))
             {
                 _oKeyMap[i].erase(kIt);
