+++ Source/System/State/Base/WS/OSGTextureObjChunk.cpp
@@ -630,7 +630,7 @@ void TextureObjChunk::handleTexture(Window                  *win,
         if(needMipmaps)
         {
             // do we have usable mipmaps ?
-            if(img->getMipMapCount() == img->calcMipmapLevelCount() &&
+            if(img->getMipMapCount() == Int32(img->calcMipmapLevelCount()) &&
                  osgIsPower2(width) && osgIsPower2(height) &&
                  osgIsPower2(depth)
               )
@@ -1248,7 +1248,7 @@ void TextureObjChunk::handleTexture(Window                  *win,
             h = ay - iy + 1;
             d = az - iz + 1;
 
-            if(w != img->getWidth())
+            if(w != UInt32(img->getWidth()))
                 glPixelStorei(GL_UNPACK_ROW_LENGTH,  img->getWidth());
             if(ix != 0)
                 glPixelStorei(GL_UNPACK_SKIP_PIXELS, ix);
@@ -1339,7 +1339,7 @@ void TextureObjChunk::handleTexture(Window                  *win,
             }
 
 
-            if(w != img->getWidth())
+            if(w != UInt32(img->getWidth()))
                 glPixelStorei(GL_UNPACK_ROW_LENGTH,  0);
             if(ix != 0)
                 glPixelStorei(GL_UNPACK_SKIP_PIXELS, 0);
@@ -1536,7 +1536,7 @@ void TextureObjChunk::activate(DrawEnv *pEnv, UInt32 idx)
     {
 #ifdef OSG_DEBUG
         FWARNING(("TextureObjChunk::activate: Trying to bind image unit %d,"
-                  " but Window %p only supports %d!\n",
+                  " but Window %p only supports %lf!\n",
                   idx, win, nteximages));
 #endif
         return;
@@ -1846,7 +1846,7 @@ void TextureObjChunk::deactivate(DrawEnv *pEnv, UInt32 idx)
     {
 #ifdef OSG_DEBUG
         FWARNING(("TextureObjChunk::deactivate: Trying to bind image unit %d,"
-                  " but Window %p only supports %d!\n",
+                  " but Window %p only supports %lf!\n",
                   idx, win, nteximages));
 #endif
         return;
